functional-annotation/requirements.txt
biopython
gffutils
